Move application-search tests to Playwright
